// packages/filter-predicates/src/predicates/types.ts
export type UnknownFilterPredicateOperator = {
  [KOperator in `$${string}`]: JsonValue;
} & {
  [KOperator in '$all' | '$any' | '$not']: never;
};
// ...
export type FilterPredicate =
  | FilterPredicateExpression
  | FilterPredicatePrimitive
  | {
      /**
       * Asserts that all of the given predicates must be true.
       */
      $all: FilterPredicate[];
    }
  | {
      /**
       * Asserts that at least one of the given predicates must be true.
       */
      $any: FilterPredicate[];
    }
  | {
      /**
       * Asserts that the given predicate must not be true.
       */
      $not: FilterPredicate;
    }
  | UnknownFilterPredicateOperator;
// ...
export type FilterPredicateExpression = {
  [KPath in string]: FilterPredicateValue;
} & {
  [KPath in `$${string}`]: never;
};

/**
 * Represents future additions to the set of filter predicate value matchers.
 *
 * @remarks
 *
 * If you write code that explicitly inspects filter predicate expressions, you
 * should be ready for the appearance of such matchers and deliberately
 * gracefully fail to match them.
 *
 * @public
 */
export type UnknownFilterPredicateValueMatcher = {
  [KMatcher in `$${string}`]: JsonValue;
} & {
  [KMatcher in '$exists' | '$in' | '$contains' | '$hasPrefix']: never;
};

/**
 * A filter predicate value that can be used to match against a property value.
 *
 * @public
 */
export type FilterPredicateValue =
  | FilterPredicatePrimitive
  | {
      /**
       * Asserts that the property exists and has any value (`true`) - or not
       * (`false`).
       */
      $exists: boolean;
    }
  | {
      /**
       * Asserts that the property value is any one of the given possible
       * values.
       */
      $in: FilterPredicatePrimitive[];
    }
  | {
      /**
       * Asserts that the property value is an array, and that at least one of
       * its elements matches the given predicate.
       */
      $contains: FilterPredicate;
    }
  | {
      /**
       * Asserts that the property value is string, and that it starts with the given string.
       */
      $hasPrefix: string;
    }
  | UnknownFilterPredicateValueMatcher;

/**
 * A primitive value that can be used in filter predicates.
 *
 * @public
 */
export type FilterPredicatePrimitive = string | number | boolean;